requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
attrs==25.3.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
croniter==6.0.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.2
fastapi-cli==0.0.12
fastapi-cloud-cli==0.2.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.41.0
googleapis-common-protos==1.70.0
GPUtil==1.4.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-google-genai==2.1.12
langchain-huggingface==0.3.1
langchain-mongodb==0.7.0
langchain-openai==0.3.33
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-checkpoint-mongodb==0.2.0
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.30
lark==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mem0ai==0.1.117
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
narwhals==2.5.0
networkx==3.5
numpy==2.3.3
ollama==0.5.4
openai==1.99.9
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
portalocker==3.2.0
posthog==6.7.6
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.14.1
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
qdrant-client==1.15.1
redis==6.4.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rq==2.6.0
rq-win==0.4.2
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
times==0.7
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
types-python-dateutil==2.9.0.20250822
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
